+++ src/nasal_builtin.h
@@ -109,11 +109,6 @@ var builtin_arch(var*, gc&);
 std::string tohex(u32);
 std::string md5(const std::string&);
 var builtin_md5(var*, gc&);
-var builtin_cocreate(var*, gc&);
-var builtin_coresume(var*, gc&);
-var builtin_coyield(var*, gc&);
-var builtin_costatus(var*, gc&);
-var builtin_corun(var*, gc&);
 var builtin_millisec(var*, gc&);
 var builtin_gcextend(var*, gc&);
 var builtin_gcinfo(var*, gc&);
